feat(task-003): extract code generation, review, and reflection pipeline
- Create CodegenPipeline in ovagent/codegen_pipeline.py owning code
  extraction, direct-Python detection, code-only agentic loop,
  lightweight review, scanpy rewriting, and full reflection
- smart_agent.py delegates all codegen methods via lazy _codegen property
  (supports __new__-constructed test doubles)
- Add code-only mode state (_code_only_mode, _code_only_captured_code,
  _code_only_captured_history) to AgentContext protocol
- Add _run_agentic_loop, _build_agentic_system_prompt, and
  _normalize_registry_entry_for_codegen to AgentContext protocol
- Remove ~1000 lines of inline implementation from smart_agent.py
- 25 new tests in test_codegen_pipeline.py covering ownership,
  delegation, and claw.py compatibility
- All 7 existing test_smart_agent.py tests continue to pass

Acceptance criteria: code generation path owned by dedicated module;
smart_agent.py delegates; code-only and execute_code semantics preserved
for claw.py consumers.
