+++ omicverse/utils/ovagent/protocol.py
@@ -1,9 +1,10 @@
 """AgentContext protocol — type-only contract for extracted ovagent modules.
 
 All extracted modules (PromptBuilder, AnalysisExecutor, ToolRuntime,
-SubagentController, TurnController) depend on this protocol instead of
-importing the concrete ``OmicVerseAgent`` class.  This breaks the circular
-import between ``smart_agent.py`` and the ``ovagent/`` subpackage.
+SubagentController, TurnController, CodegenPipeline) depend on this
+protocol instead of importing the concrete ``OmicVerseAgent`` class.
+This breaks the circular import between ``smart_agent.py`` and the
+``ovagent/`` subpackage.
 
 The protocol is **not** ``@runtime_checkable`` — it is purely a
 type-checking aid (no runtime overhead).
@@ -67,6 +68,11 @@ class AgentContext(Protocol):
     _web_session_id: str
     _managed_api_env: Dict[str, str]
 
+    # ---- code-only mode state (used by CodegenPipeline and ToolRuntime) ----
+    _code_only_mode: bool
+    _code_only_captured_code: str
+    _code_only_captured_history: List[Dict[str, Any]]
+
     # ---- feature flags ----
     use_notebook_execution: bool
     enable_filesystem_context: bool
@@ -111,3 +117,15 @@ class AgentContext(Protocol):
     def _load_skill_guidance(self, slug: str) -> str: ...
 
     def _extract_python_code(self, text: str) -> Optional[str]: ...
+
+    def _normalize_registry_entry_for_codegen(self, entry: Dict[str, Any]) -> Dict[str, Any]: ...
+
+    def _build_agentic_system_prompt(self) -> str: ...
+
+    async def _run_agentic_loop(
+        self, request: str, adata: Any,
+        event_callback: Any = None,
+        cancel_event: Any = None,
+        history: Any = None,
+        approval_handler: Any = None,
+    ) -> Any: ...
